pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "armarius"
version = "0.1.0"
description = "Cryptographic guardian for AI agents — architectural prevention of prompt injection"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "PyNaCl>=1.5.0",
]
keywords = [
    "security",
    "ai",
    "prompt-injection",
    "agents",
    "cryptography",
    "llm",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

[project.urls]
Homepage = "https://github.com/tatlantis/armarius"
Repository = "https://github.com/tatlantis/armarius"
Issues = "https://github.com/tatlantis/armarius/issues"

[tool.setuptools.packages.find]
include = ["armarius*"]
